Techstack: Fix the issue with pin.
Before we were doing the horizontal scroll by manually fixing the
techstack section from the onEnter function. But I was causing some
issue with the divs placed below. But now we are using pin property
instead of manually fixing the section.
Previously pin property was not used because it was making the section
disappear when the animation starts but that was do to some react and
gsap cleanup issues. Now we are using gsap context to make sure the
cleanup is handled properly and it solves the pinning issue.